Update dev scripts
- Update all dev scripts with small improvments
- Install kitty if OS is Arch, and Terminator if OS is Ubuntu
- Install fastfetch instead of neofetch

+++ new-code/dev/update-code-on-vm.sh
@@ -1,17 +1,35 @@
 #!/bin/bash
 
+os_name=$(grep "^NAME=" /etc/os-release | cut -d '=' -f2 | sed 's/"//g')
+
+case $os_name in
+  Ubuntu) os_color="#E95420" ;;
+  Arch) os_color="#1793D1" ;;
+  *)
+    print_message "Unsupported OS"
+    exit 1
+    ;;
+esac
+
 vm=$(gum choose \
-		--cursor.foreground="#E95420" \
-		--selected.foreground="#E95420" \
+        --header "What the OS?" \
+        --header.foreground="#FFFFFF" \
+		--cursor.foreground="$os_color" \
+		--selected.foreground="$os_color" \
 		--no-show-help \
 		"ubuntu" "arch")
 
-if [ "$vm" == "ubuntu" ];then
-    vm_name="ubuntu-dotfiles-vm"
-else
-    echo "$vm not supported yet"
+if [ -z "$vm" ];then
     exit 0
 fi
 
-vm_ip=$(incus list "$vm_name" -f json | jq -r '.[].state.network.enp5s0.addresses[0].address')
-rsync --delete -avz . "ubuntu@${vm_ip}:dotfiles/"
+vm_name="$vm-dotfiles-vm"
+vm_search=$(incus list "$vm_name" -f json | jq -r '.[].name')
+
+if [[ -n "$vm_search" ]];then
+	vm_ip=$(incus list "$vm_name" -f json | jq -r '.[].state.network.enp5s0.addresses[0].address')
+	rsync --delete -avz . "ubuntu@${vm_ip}:dotfiles/"
+else
+	echo "$vm doens't exists"
+	exit 0
+fi
